meta-qti-telaf: Add restorecon for telaf ubiblock
Runs restorecon after ubbiblock is created in mount script.
This is required as a new sepolicy has been added for verified_boot.
Once the ubiblock is created it picks up the default context for
/dev/ubiblock*. Since the ubiblock is read in the verified_boot
util, vb fails with avc denials.
The restorecon sets the correct selinux context after the
node is created.

Change-Id: I8ea11e3aef8d9e7f26af3d0ae697457be45fc0a3