meta-qti-telaf: Fix TelAF boot failed issue
The mini NAD2 has no audio service, so the dependency
of "init_audio.service" can not work, this make the
telaf service boot up before the basic system resource
was ready. The cannot lock the resource and trigger
reboot.

Make the TelAF boot after "basic.target" to fix this
issue.

Change-Id: I68a70873015780c65d18d09f181ec69891b7219d